project/requirements.txt
alibabacloud_dysmsapi20170525==2.0.23
alibabacloud_tea_openapi==0.3.7
alibabacloud_tea_util==0.3.8
Faker==18.10.0
Flask==2.2.3
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
Flask_SocketIO==5.3.4
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.1.1
pycryptodome==3.17
PyJWT==2.6.0
pytest==7.3.1
qiniu==7.10.0
Requests==2.31.0
SQLAlchemy==2.0.9
Werkzeug==2.2.3